[ft-164198489] refactor methods and filename
