+++ package.json
@@ -46,10 +46,10 @@
     "sequelize": "^4.42.0",
     "sequelize-cli": "^5.4.0",
     "slug": "^0.9.1",
-    "underscore": "^1.9.1"
+    "underscore": "^1.9.1",
+    "@babel/polyfill": "^7.2.5"
   },
   "devDependencies": {
-    "@babel/polyfill": "^7.2.5",
     "@babel/preset-env": "^7.3.4",
     "@babel/register": "^7.0.0",
     "coveralls": "^3.0.3",
